refactor to update/utils.ts
