+++ packages/eas-cli/src/commands/update/view.ts
@@ -1,30 +1,11 @@
 import { Command, flags } from '@oclif/command';
 import Table from 'cli-table3';
 import gql from 'graphql-tag';
-import groupBy from 'lodash/groupBy';
-import { format } from 'timeago.js';
 
 import { graphqlClient, withErrorHandlingAsync } from '../../graphql/client';
-import {
-  Maybe,
-  Robot,
-  Update,
-  UpdatesByGroupQuery,
-  UpdatesByGroupQueryVariables,
-  User,
-} from '../../graphql/generated';
+import { UpdatesByGroupQuery, UpdatesByGroupQueryVariables } from '../../graphql/generated';
 import Log from '../../log';
-import { getActorDisplayName } from '../../user/User';
-export const UPDATE_COLUMNS = [
-  'update description',
-  'update runtime version',
-  'update group ID',
-  'platforms',
-];
-
-export type FormatUpdateParameter = Pick<Update, 'id' | 'createdAt' | 'message'> & {
-  actor?: Maybe<Pick<User, 'username' | 'id'> | Pick<Robot, 'firstName' | 'id'>>;
-};
+import { UPDATE_COLUMNS, formatUpdate, getPlatformsForGroup } from '../../update/utils';
 
 export async function viewUpdateAsync({
   groupId,
@@ -110,31 +91,10 @@ export default class UpdateView extends Command {
       representativeUpdate.group,
       getPlatformsForGroup({
         updates: updatesByGroup,
-        group: updatesByGroup[0].group,
+        group: updatesByGroup[0]?.group,
       }),
     ]);
 
     Log.log(groupTable.toString());
   }
-}
-
-export function getPlatformsForGroup({
-  group,
-  updates,
-}: {
-  group: string;
-  updates: { group: string; platform: string }[];
-}): string {
-  const groupedUpdates = groupBy(updates, update => update.group);
-  return groupedUpdates[group].map(update => update.platform).join(',');
-}
-
-export function formatUpdate(update: FormatUpdateParameter): string {
-  if (!update) {
-    return 'N/A';
-  }
-  const message = update.message ? `"${update.message}" ` : '';
-  return `${message}(${format(update.createdAt, 'en_US')} by ${getActorDisplayName(
-    update.actor as any
-  )})`;
 }
