+++ packages/eas-cli/src/update/utils.ts
@@ -1,0 +1,43 @@
+import groupBy from 'lodash/groupBy';
+import { format } from 'timeago.js';
+
+import { Maybe, Robot, Update, User } from '../graphql/generated';
+import { getActorDisplayName } from '../user/User';
+
+export type FormatUpdateParameter = Pick<Update, 'id' | 'createdAt' | 'message'> & {
+  actor?: Maybe<Pick<User, 'username' | 'id'> | Pick<Robot, 'firstName' | 'id'>>;
+};
+
+export const UPDATE_COLUMNS = [
+  'update description',
+  'update runtime version',
+  'update group ID',
+  'platforms',
+];
+
+export function getPlatformsForGroup({
+  group,
+  updates,
+}: {
+  group: string;
+  updates: { group: string; platform: string }[];
+}): string {
+  const groupedUpdates = groupBy(updates, update => update.group);
+  if (Object.keys(groupedUpdates).length === 0) {
+    return 'N/A';
+  }
+  return groupedUpdates[group]
+    .map(update => update.platform)
+    .sort()
+    .join(',');
+}
+
+export function formatUpdate(update: FormatUpdateParameter): string {
+  if (!update) {
+    return 'N/A';
+  }
+  const message = update.message ? `"${update.message}" ` : '';
+  return `${message}(${format(update.createdAt, 'en_US')} by ${getActorDisplayName(
+    update.actor as any
+  )})`;
+}
